+++ app/src/main/cpp/hideload/linker_debug.h
@@ -7,7 +7,7 @@
 #include <android/log.h>
 
 
-#define LOG_TAG "CustomLinker"
+#define LOG_TAG "hideLinker"
 #define LOGE(...) __android_log_print(ANDROID_LOG_ERROR,LOG_TAG,__VA_ARGS__)
 
 #define TRACECODE(...) __android_log_print(ANDROID_LOG_ERROR,LOG_TAG,"[file %s],[line %d],[function:%s]",__FILE__, __LINE__,__func__);
